tighten EXPORT/LOCAL decls, visbility, fix linking
exp. for Windows, and less .so symbols.
we cannot link to the so with so many private symbols, at least not
on Windows, so fixup the visibility. link u8idlint to the sources
(the trick is to specify u8idlint_CFLAGS = $(AM_CFLAGS) to create
seperate -fPIC libtool and static objects), not to the lib.